+++ sdk/azurestackhci/Azure.ResourceManager.Hci/CHANGELOG.md
@@ -1,6 +1,13 @@
 # Release History
 
-## 1.3.0-beta.1 (Unreleased)
+## 1.3.0-beta.2 (2026-04-24)
+
+### Other Changes
+
+- Upgraded dependent `Azure.Core` to 1.53.0.
+- Upgraded dependent `Azure.ResourceManager` to 1.14.0.
+
+## 1.3.0-beta.1 (2026-04-23)
 
 ### Features Added
 
@@ -14,8 +21,6 @@
 - Some property types changed (e.g., `ConnectivityProperties` from `BinaryData` to a typed model).
 - Some resource collection and extension method signatures changed.
 - Some model factory overloads changed signatures.
-
-### Bugs Fixed
 
 ### Other Changes